Lots of improvements for clang

+++ Devices/GPS/Sim/GpsSimulator.cpp
@@ -28,10 +28,11 @@ namespace device {
 
 
             GpsMeasurement_t gps{};
-            auto noiseLat = (noiseDist(rng) * sigmaLat) / EARTH_CIRCUMFERENCE * 360;
+            auto noiseLat = (noiseDist(rng) * sigmaLat) / static_cast<float>(EARTH_CIRCUMFERENCE) * 360;
             gps.location.lat = 48.4235495 + noiseLat;
             auto noiseLon =
-                    (noiseDist(rng) * sigmaLon) / EARTH_CIRCUMFERENCE * 360 / std::cos(gps.location.lat / 180 * M_PI);
+                    (noiseDist(rng) * sigmaLon) /
+                            static_cast<float>(EARTH_CIRCUMFERENCE) * 360 / std::cos(gps.location.lat / 180 * M_PI);
             gps.location.lon = 9.9292937 + noiseLon;
             gps.timestamp = si::Second<long double>{0};
             gps.location.altitude = 500_meter + noiseDist(rng) * sigmaVert * meter;
